StandAlone-Bootstrap-Page
A stand alone page application of Bootstrap without CSS styles. It is to
display the UI with bootstrap classes, which are much easier to use and
understand.
Also this application suggest the use of Glyphycon, tabs, wells,
dropdowns and jummbotron. Based on the Mobile-First Technology, it has
unmovable panels and
formated structure.